requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
rich>=13.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
